Move out prisonInformation function to supportingInformationController
Now we have a controller for supportingInformation it makes sense
to move this method into it
