+++ server/controllers/assess/assessmentsController.test.ts
@@ -1,27 +1,17 @@
 import type { Request, Response, NextFunction } from 'express'
 import type { GroupedAssessments } from '@approved-premises/ui'
 import { createMock, DeepMocked } from '@golevelup/ts-jest'
-import { ApprovedPremisesAssessment as Assessment } from '../../@types/shared'
 
 import TasklistService from '../../services/tasklistService'
 import AssessmentsController, { tasklistPageHeading } from './assessmentsController'
 import { AssessmentService } from '../../services'
 
 import assessmentFactory from '../../testutils/factories/assessment'
-import adjudicationFactory from '../../testutils/factories/adjudication'
-import prisonCaseNotesFactory from '../../testutils/factories/prisonCaseNotes'
 
 import paths from '../../paths/assess'
 import informationSetAsNotReceived from '../../utils/assessments/informationSetAsNotReceived'
 import getSections from '../../utils/assessments/getSections'
-import { DateFormats } from '../../utils/dateUtils'
-import acctAlertFactory from '../../testutils/factories/acctAlert'
-import {
-  acctAlertsFromAssessment,
-  adjudicationsFromAssessment,
-  caseNotesFromAssessment,
-  groupAssessmements,
-} from '../../utils/assessments/utils'
+import { groupAssessmements } from '../../utils/assessments/utils'
 import { hasRole } from '../../utils/userUtils'
 
 jest.mock('../../utils/assessments/utils')
@@ -242,41 +232,6 @@ describe('assessmentsController', () => {
           assessment,
         })
       })
-    })
-  })
-
-  describe('prisonInformation', () => {
-    let assessment: Assessment
-
-    beforeEach(() => {
-      request.params.id = 'some-id'
-      assessment = assessmentFactory.build()
-
-      assessmentService.findAssessment.mockResolvedValue(assessment)
-    })
-
-    it('renders the view', async () => {
-      const adjudications = adjudicationFactory.buildList(2)
-      const caseNotes = prisonCaseNotesFactory.buildList(2)
-      const acctAlerts = acctAlertFactory.buildList(2)
-
-      ;(adjudicationsFromAssessment as jest.Mock).mockReturnValue(adjudications)
-      ;(caseNotesFromAssessment as jest.Mock).mockReturnValue(caseNotes)
-      ;(acctAlertsFromAssessment as jest.Mock).mockReturnValue(acctAlerts)
-
-      const requestHandler = assessmentsController.prisonInformation()
-
-      await requestHandler(request, response, next)
-
-      expect(response.render).toBeCalledWith('assessments/pages/risk-information/prison-information', {
-        adjudications,
-        caseNotes,
-        acctAlerts,
-        assessmentId: assessment.id,
-        dateOfImport: DateFormats.isoDateToUIDate(assessment.application.submittedAt),
-        pageHeading: 'Prison information',
-      })
-      expect(assessmentService.findAssessment).toBeCalledWith(token, request.params.id)
     })
   })
 })
