+++ server/paths/assess.ts
@@ -6,7 +6,6 @@ const assessmentPath = assessmentsPath.path(':id')
 
 const pagesPath = assessmentPath.path('tasks/:task/pages/:page')
 
-const prisonInformationPath = assessmentPath.path('prison-information')
 const supportingInformationPath = assessmentPath.path('supporting-information/:category')
 
 const submission = assessmentPath.path('submission')
@@ -18,7 +17,6 @@ const paths = {
     pages: {
       show: pagesPath,
       update: pagesPath,
-      prisonInformationPath,
     },
     supportingInformationPath,
     clarificationNotes: {
